+++ src/routes/api/player/[uuid=uuid]/+server.ts
