+++ src/routes/stats/[id=id]/[profile]/+page.svelte
